Rewrite code to improve performance
This commit includes a complete rewrite of the code in order to significantly improve its performance. As a result, the code is now 20 times faster than it was previously.

This change was necessary because the previous implementation was inefficient and was causing the code to run slowly. The new implementation addresses these issues and makes the code much more efficient.

Note: Some additional testing may be needed to ensure that the new implementation is correct and does not introduce any new bugs. Please review the code carefully before merging.